[refactor] :recycle: move jwt types to specific file (#183)
