+++ src/app/api/v1/user/[id]/route.ts
@@ -1,14 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server'
 
 import { getIpFromRequest } from 'src/app/api/_utils/request-utils'
-import { JwtPayload, extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
+import { extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
 import {
   updateUser,
   getUserById,
-  getUserByPhone,
   checkUserHaveActiveSession
 } from 'src/app/api/_data/data-service'
 
+import { JwtPayload } from 'src/types/jwt'
 import { IAccount } from 'src/types/account'
 
 // ----------------------------------------------------------------------
