+++ src/app/api/v1/user/[id]/transfer-all/route.ts
@@ -2,9 +2,10 @@ import { NextRequest, NextResponse } from 'next/server'
 
 import { transferAll } from 'src/app/api/_data/blk-service'
 import { getIpFromRequest } from 'src/app/api/_utils/request-utils'
-import { JwtPayload, extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
+import { extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
 import { getUserById, checkUserHaveActiveSession } from 'src/app/api/_data/data-service'
 
+import { JwtPayload } from 'src/types/jwt'
 import { IAccount } from 'src/types/account'
 
 // ----------------------------------------------------------------------
