+++ src/app/api/v1/wallet/[id]/transactions/route.ts
@@ -1,13 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server'
 
 import { getIpFromRequest } from 'src/app/api/_utils/request-utils'
-import { JwtPayload, extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
+import { extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
 import {
   getUserIdByWallet,
   getUserTransactions,
   checkUserHaveActiveSession
 } from 'src/app/api/_data/data-service'
 
+import { JwtPayload } from 'src/types/jwt'
 import { IErrorResponse } from 'src/types/api'
 import { ITransaction } from 'src/types/wallet'
