+++ src/auth/context/jwt/auth-provider.tsx
@@ -2,8 +2,12 @@
 
 import { useMemo, useEffect, useReducer, useCallback } from 'react'
 
+import { getStorageItem } from 'src/hooks/use-local-storage'
+
 import { STORAGE_KEY_TOKEN } from 'src/config-global'
 import { post, fetcher, endpoints } from 'src/app/api/_hooks/api-resolver'
+
+import { JwtPayload, jwtPayloadUser } from 'src/types/jwt'
 
 import { AuthContext } from './auth-context'
 import { jwtDecode, setSession, isValidToken, getAuthorizationHeader } from './utils'
